remember me

+++ lib/core/services/storage/local_storage_service.dart
@@ -11,10 +11,36 @@ final class LocalStorageService implements ILocalStorageService {
   factory LocalStorageService() => _instance;
   LocalStorageService._internal();
 
-  Future<void> init() async => _prefs = await SharedPreferences.getInstance();
+  Future<void> init() async {
+    _prefs = await SharedPreferences.getInstance();
+    _isSessionMode = !(await readKey(StorageKey.rememberMe) ?? false);
+  }
+
+  /// Session mode flag
+  bool _isSessionMode = false;
+
+  /// In-memory cache (cleared when app is killed)
+  final Map<String, dynamic> _sessionCache = {};
+
+  @override
+  void disableSessionMode(bool value) {
+    _isSessionMode = value;
+
+    if (!value) {
+      _sessionCache.clear();
+    }
+  }
+
+  bool get isSessionMode => _isSessionMode;
 
   @override
   Future<void> saveKey(StorageKey key, dynamic value) async {
+    /// If session mode → save only in memory
+    if (_isSessionMode) {
+      _sessionCache[key.key] = value;
+      return;
+    }
+
     switch (key.type) {
       case StorageKeyType.string:
         await _prefs.setString(key.key, value as String);
@@ -33,6 +59,10 @@ final class LocalStorageService implements ILocalStorageService {
 
   @override
   Future<dynamic> readKey(StorageKey key) async {
+    /// If session mode → read from memory
+    if (_isSessionMode) {
+      return _sessionCache[key.key];
+    }
     switch (key.type) {
       case StorageKeyType.string:
         return _prefs.getString(key.key);
@@ -47,6 +77,12 @@ final class LocalStorageService implements ILocalStorageService {
 
   @override
   Future<void> deleteKey(StorageKey key) async {
+    /// Session mode
+    if (_isSessionMode) {
+      _sessionCache.remove(key.key);
+      return;
+    }
+
     if (key.type == StorageKeyType.secureString) {
       await _secureStorage.delete(key: key.key);
     } else {
@@ -55,13 +91,24 @@ final class LocalStorageService implements ILocalStorageService {
   }
 
   @override
-  Future<void> clearPrefs() async => await _prefs.clear();
+  Future<void> clearPrefs() async {
+    _sessionCache.clear();
+
+    await _prefs.clear();
+  }
 
   @override
-  Future<void> clearSecure() async => await _secureStorage.deleteAll();
+  Future<void> clearSecure() async {
+    _sessionCache.clear();
+
+    await _secureStorage.deleteAll();
+  }
 
   @override
   Future<void> clearAll() async {
+    _sessionCache.clear();
+    _isSessionMode = true;
+
     await clearPrefs();
     await clearSecure();
   }
